backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
requests==2.32.3
psycopg2-binary==2.9.9
qdrant-client==1.9.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
tiktoken==0.7.0
